Split hardware tests into separate files and add socket optimizations
