+++ sdkconfig.defaults
@@ -16,8 +16,10 @@ CONFIG_HTTPD_WS_SUPPORT=y
 CONFIG_HTTPD_MAX_REQ_HDR_LEN=1024
 CONFIG_HTTPD_MAX_URI_LEN=512
 
-# lwIP - increase socket count for relay
-CONFIG_LWIP_MAX_SOCKETS=32
+# lwIP - optimize for relay workload
+CONFIG_LWIP_MAX_SOCKETS=48
+CONFIG_LWIP_SO_REUSE=y
+CONFIG_LWIP_SO_LINGER=y
 CONFIG_LWIP_TCP_MSS=1436
 CONFIG_LWIP_TCP_SND_BUF_DEFAULT=11456
 CONFIG_LWIP_TCP_WND_DEFAULT=11456
